removed unused codes. fix a bug about remote run. (#13275)
